requirements.txt
torch>=1.10.0
torchtext>=0.10.0
torchvision>=0.11.1
tokenizers>=0.10.0
numpy>=1.21.0
Pillow >= 2.1.0
transformers
Flask
requests